refactor(#29) pathvariable 추가, 피드백 반영

+++ src/main/java/swith/swithServer/domain/studyGroup/dto/GroupResponse.java
@@ -14,8 +14,8 @@ import lombok.Builder;
 public class GroupResponse {
     private String groupName;
     private String groupInsertId;
-    private int maxNum;
+    private Long maxNum;
     private String subject;
-    private int period;
+    private Long period;
     private String communication;
 }
